Commiting for UspSearchCustomerandCardForm Changed Form Number Datatype

+++ HPCL.DataModel/Customer/SearchCustomerandCardFormModel.cs
@@ -57,7 +57,7 @@ namespace HPCL.DataModel.Customer
 
         [JsonProperty("FormNumber")]
         [DataMember]
-        public Int32 FormNumber { get; set; }
+        public Int64 FormNumber { get; set; }
 
         [JsonProperty("CustomerStatus")]
         [DataMember]
@@ -104,7 +104,7 @@ namespace HPCL.DataModel.Customer
 
         [JsonProperty("FormNumber")]
         [DataMember]
-        public Int32 FormNumber { get; set; }
+        public Int64 FormNumber { get; set; }
 
         [JsonProperty("Status")]
         [DataMember]
